pyproject.toml
[project]
name = "musicmind-mcp"
version = "0.1.0"
description = "MCP server for intelligent Apple Music integration with taste profiling"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "mcp>=1.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "sqlalchemy>=2.0",
    "pyjwt[crypto]>=2.8",
    "cryptography>=42.0",
    "numpy>=1.26",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/musicmind"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
